services/backend/requirements.txt
aerich
asyncpg
fastapi
tortoise-orm
uvicorn
fastapi-mail
docxtpl
docx2pdf